requirements.txt
asgiref==3.8.1
black==24.4.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.7
Django==4.2.13
djangorestframework==3.15.1
gunicorn==22.0.0
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
pdfminer.six==20231228
pdfplumber==0.11.0
pillow==10.3.0
platformdirs==4.2.2
pycparser==2.22
pypdfium2==4.30.0
sqlparse==0.5.0
tomli==2.0.1
typing_extensions==4.12.0
